+++ src/utils/warningManager.js
@@ -1,0 +1,7 @@
+const showWarning = data => {
+  console.warn(data);
+};
+
+export default {
+  showWarning
+};